+++ templates/demo/sandbox/form-inputs/form-inputs.js
@@ -47,11 +47,9 @@ class FormInputs extends React.Component {
   render () {
     return <>
       <link rel="stylesheet" href="https://www.ft.com/__origami/service/build/v2/bundles/css?modules=o-forms@styles&brand=internal" />
-      {
-        Object.entries(this.props.data.config).map(([type, config]) => this.buildInputType(config))
-        // TODO: handle error
-      }
+      { Object.entries(this.props.data.config).map(([type, config]) => this.buildInputType(config)) }
     </>
+      // TODO: handle error
   }
 
 }
